feat(classloader): add type checking
- This makes it fail with a proper error message
- Needed some reshuffling of modules to avoid circular imports

Fixes #11976
