+++ weblate/addons/base.py
@@ -16,11 +16,8 @@ from django.utils.functional import cached_property
 from django.utils.translation import gettext
 
 from weblate.addons.events import AddonEvent
-from weblate.addons.forms import BaseAddonForm
-from weblate.addons.tasks import postconfigure_addon
 from weblate.trans.exceptions import FileParseError
 from weblate.trans.models import Component
-from weblate.trans.tasks import perform_update
 from weblate.trans.util import get_clean_env
 from weblate.utils import messages
 from weblate.utils.errors import report_error
@@ -30,6 +27,7 @@ from weblate.utils.validators import validate_filename
 if TYPE_CHECKING:
     from django_stubs_ext import StrOrPromise
 
+    from weblate.addons.forms import BaseAddonForm
     from weblate.addons.models import Addon
     from weblate.auth.models import User
     from weblate.formats.base import TranslationFormat
@@ -154,6 +152,8 @@ class BaseAddon:
         self.post_configure()
 
     def post_configure(self, run: bool = True) -> None:
+        from weblate.addons.tasks import postconfigure_addon
+
         self.instance.log_debug("configuring events for %s add-on", self.name)
 
         # Configure events to current status
@@ -374,6 +374,8 @@ class BaseAddon:
 
     @classmethod
     def pre_install(cls, obj: Component | Project | None, request) -> None:
+        from weblate.trans.tasks import perform_update
+
         if cls.trigger_update and isinstance(obj, Component):
             perform_update.delay("Component", obj.pk, auto=True)
             if obj.repo_needs_merge():
@@ -398,15 +400,6 @@ class BaseAddon:
         return User.objects.get_or_create_bot(
             "addon", self.user_name, self.user_verbose
         )
-
-
-class TestAddon(BaseAddon):
-    """Testing add-on doing nothing."""
-
-    settings_form = BaseAddonForm
-    name = "weblate.base.test"
-    verbose = "Test add-on"
-    description = "Test add-on"
 
 
 class UpdateBaseAddon(BaseAddon):
@@ -434,26 +427,6 @@ class UpdateBaseAddon(BaseAddon):
         with suppress(FileParseError):
             self.update_translations(component, previous_head)
         self.commit_and_push(component, skip_push=skip_push)
-
-
-class TestError(Exception):
-    pass
-
-
-class TestCrashAddon(UpdateBaseAddon):
-    """Testing add-on doing nothing."""
-
-    name = "weblate.base.crash"
-    verbose = "Crash test add-on"
-    description = "Crash test add-on"
-
-    def update_translations(self, component: Component, previous_head: str) -> None:
-        if previous_head:
-            raise TestError("Test error")
-
-    @classmethod
-    def can_install(cls, component: Component, user: User | None) -> bool:  # noqa: ARG003
-        return False
 
 
 class StoreBaseAddon(BaseAddon):
